abstraction and remove testing
